Fix: Error handring for tar command.

+++ jenkins-backup.sh
@@ -74,6 +74,11 @@ fi
 
 cd "$TMP_DIR"
 tar -czvf "$TMP_TAR_NAME" "$ARC_NAME/"*
+exitcode=$?
+if [ "$exitcode" != "1" ] && [ "$exitcode" != "0" ]; then
+    echo "Error!" >&2
+    exit $exitcode
+fi
 cd -
 mv -f "$TMP_TAR_NAME" "$DEST_FILE"
 rm -rf "$ARC_DIR"
